+++ apps/dash-medical-provider-charges/requirements.txt
@@ -1,6 +1,5 @@
-dash==1.12.0
-gunicorn>=19.9.0
-numpy==1.16.2
-pandas==0.24.2
+dash==2.4.1
+pandas==1.4.2
+gunicorn==20.1.0
 uszipcode==0.2.2
-plotly==4.7.1
+
